requirements.txt
allure-pytest==2.13.5
Faker==24.14.0
pytest==8.1.1
pytest-xdist==3.5.0
selene==2.0.0rc9
selenium==4.20.0